Update BresserWeatherSensorOptions.ino
Added typecast for RP2040

+++ examples/BresserWeatherSensorOptions/BresserWeatherSensorOptions.ino
@@ -60,7 +60,7 @@ void print_data(void)
       continue;
 
     Serial.printf("Id: [%8X] Typ: [%X] Battery: [%s] ",
-        weatherSensor.sensor[i].sensor_id,
+        (unsigned int)weatherSensor.sensor[i].sensor_id,
         weatherSensor.sensor[i].s_type,
         weatherSensor.sensor[i].battery_ok ? "OK " : "Low");
     #ifdef BRESSER_6_IN_1
